+++ misc/multistream-select/src/lib.rs
@@ -87,6 +87,8 @@
 //! ```
 //!
 
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+
 mod dialer_select;
 mod length_delimited;
 mod listener_select;