+++ protocols/mdns/src/lib.rs
@@ -30,6 +30,9 @@
 //! implements the `NetworkBehaviour` trait. This struct will automatically discover other
 //! libp2p nodes on the local network.
 //!
+
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+
 use std::net::{Ipv4Addr, Ipv6Addr};
 use std::time::Duration;