+++ swarm/CHANGELOG.md
@@ -1,4 +1,8 @@
-# 0.40.0
+# 0.41.0 [unreleased]
+
+- Update to `libp2p-core` `v0.38.0`.
+
+# 0.40.1
 
 - Bump rand to 0.8 and quickcheck to 1. See [PR 2857].